mcp/python/pyproject.toml
[project]
name = "clinical-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-exporter-otlp-proto-grpc>=1.24",
  "jsonschema>=4.22"
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.5", "mypy>=1.5"]
